Make footer as light as possible with AA contrast
Make the footer text as light as possible while still getting a AA
rating (>= 4.5) for the color contrast. This change makes it easier to
see the text on the dark theme. The color #767676 has a contrast ratio
of 4.54 on the light background (#FFFFFF) and 3.15 on the default dark
background (#202B38).